# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация StudX API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Локальный сервер"
        }
    ],
    "paths": {
...
        "/courses/{courseId}/tasks": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Получение всех задач курса",
                "operationId": "387f13369d82d5b9f8eb0435ce6b5f6d",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "Идентификатор курса",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение всех задач курса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/tasks"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{taskId}": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Получение информации о задаче",
                "operationId": "442fc90159526ac0a98d0bc1fd59cd04",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "Идентификатор задачи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение информации о задаче",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/tasks"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Задача не найдена"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "tasks": {
                "title": "Задача",
                "description": "Схема задачи",
                "required": [
                    "id",
                    "course_id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Идентификатор задачи",
                        "type": "integer",
                        "example": 1
                    },
                    "course_id": {
                        "description": "Идентификатор курса",
                        "type": "string",
                        "example": "1"
                    },
                    "name": {
                        "description": "Название задачи",
                        "type": "string",
                        "example": "HTML & CSS Basics"
                    },
                    "created_at": {
                        "description": "Дата и время создания задачи",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-19T12:34:56Z"
                    },
                    "updated_at": {
                        "description": "Дата и время последнего обновления задачи",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-19T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}